Report collector code extracted + script to run multiple queries
