+++ src/bin/test_mpc.rs
@@ -4,34 +4,17 @@ use generic_array::ArrayLength;
 use hyper::http::uri::Scheme;
 use ipa::{
     cli::{
-        playbook::{playbook_ipa, secure_mul, InputSource},
-        CsvSerializer, Verbosity,
+        playbook::{secure_mul, InputSource},
+        Verbosity,
     },
     config::{ClientConfig, NetworkConfig, PeerConfig},
     ff::{Field, FieldType, Fp31, Fp32BitPrime, Serializable},
-    helpers::query::{IpaQueryConfig, QueryConfig, QueryType},
-    hpke::{KeyRegistry, PublicKeyOnly},
+    helpers::query::{QueryConfig, QueryType},
     net::{ClientIdentity, MpcHelperClient},
-    protocol::{BreakdownKey, MatchKey, QueryId},
-    report::{KeyIdentifier, DEFAULT_KEY_ID},
+    protocol::QueryId,
     secret_sharing::{replicated::semi_honest::AdditiveShare, IntoShares},
-    test_fixture::{
-        ipa::{ipa_in_the_clear, IpaSecurityModel, TestRawDataRecord},
-        EventGenerator, EventGeneratorConfig,
-    },
 };
-use rand::thread_rng;
-use std::{
-    error::Error,
-    fmt::Debug,
-    fs,
-    fs::OpenOptions,
-    io,
-    io::{stdout, Write},
-    ops::Add,
-    path::PathBuf,
-    time::{Duration, Instant},
-};
+use std::{error::Error, fmt::Debug, fs, ops::Add, path::PathBuf, time::Duration};
 use tokio::time::sleep;
 
 #[derive(Debug, Parser)]
@@ -89,23 +72,6 @@ impl From<&CommandInput> for InputSource {
 enum TestAction {
     /// Execute end-to-end multiplication.
     Multiply,
-    /// Execute IPA in semi-honest honest majority setting
-    SemiHonestIpa(IpaQueryConfig),
-    /// Execute IPA in malicious honest majority setting
-    MaliciousIpa(IpaQueryConfig),
-    /// Generate inputs for IPA
-    GenIpaInputs {
-        /// Number of records to generate
-        #[clap(long, short = 'n')]
-        count: u32,
-
-        /// The destination file for generated records
-        #[arg(long)]
-        output_file: Option<PathBuf>,
-
-        #[clap(flatten)]
-        gen_args: EventGeneratorConfig,
-    },
 }
 
 #[derive(Debug, clap::Args)]
@@ -212,180 +178,9 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     match args.action {
         TestAction::Multiply => multiply(&args, &make_clients().await).await,
-        TestAction::SemiHonestIpa(config) => {
-            // semi_honest_ipa(&args, &config, &make_clients().await).await
-            ipa(
-                &args,
-                &network,
-                IpaSecurityModel::SemiHonest,
-                &config,
-                &make_clients().await,
-            )
-            .await
-        }
-        TestAction::MaliciousIpa(config) => {
-            // malicious_ipa(&args, &config, &make_clients().await).await
-            ipa(
-                &args,
-                &network,
-                IpaSecurityModel::Malicious,
-                &config,
-                &make_clients().await,
-            )
-            .await
-        }
-        TestAction::GenIpaInputs {
-            count,
-            output_file,
-            gen_args,
-        } => gen_inputs(count, output_file, gen_args).unwrap(),
     };
 
     Ok(())
-}
-
-fn gen_inputs(
-    count: u32,
-    output_file: Option<PathBuf>,
-    args: EventGeneratorConfig,
-) -> io::Result<()> {
-    let event_gen = EventGenerator::with_config(thread_rng(), args).take(count as usize);
-    let mut writer: Box<dyn Write> = if let Some(path) = output_file {
-        Box::new(OpenOptions::new().write(true).create_new(true).open(path)?)
-    } else {
-        Box::new(stdout().lock())
-    };
-
-    for event in event_gen {
-        event.to_csv(&mut writer)?;
-        writer.write(&[b'\n'])?;
-    }
-
-    Ok(())
-}
-
-#[derive(Default)]
-struct KeyRegistries(Vec<KeyRegistry<PublicKeyOnly>>);
-
-impl KeyRegistries {
-    fn init_from(
-        &mut self,
-        network: &NetworkConfig,
-    ) -> Option<(KeyIdentifier, [&KeyRegistry<PublicKeyOnly>; 3])> {
-        // Get the configs, if all three peers have one
-        let Some(configs) = network
-            .peers()
-            .iter()
-            .fold(Some(vec![]), |acc, peer| {
-                if let (Some(mut vec), Some(hpke_config)) = (acc, peer.hpke_config.as_ref()) {
-                    vec.push(hpke_config);
-                    Some(vec)
-                } else {
-                    None
-                }
-            })
-        else {
-            return None;
-        };
-
-        // Create key registries
-        self.0 = configs
-            .into_iter()
-            .map(|hpke| KeyRegistry::from_keys([PublicKeyOnly(hpke.public_key.clone())]))
-            .collect::<Vec<KeyRegistry<PublicKeyOnly>>>();
-
-        Some((
-            DEFAULT_KEY_ID,
-            self.0.iter().collect::<Vec<_>>().try_into().ok().unwrap(),
-        ))
-    }
-}
-
-async fn ipa(
-    args: &Args,
-    network: &NetworkConfig,
-    security_model: IpaSecurityModel,
-    ipa_query_config: &IpaQueryConfig,
-    helper_clients: &[MpcHelperClient; 3],
-) {
-    let input = InputSource::from(&args.input);
-    let query_type: QueryType;
-    match security_model {
-        IpaSecurityModel::SemiHonest => {
-            query_type = QueryType::SemiHonestIpa(ipa_query_config.clone());
-        }
-        IpaSecurityModel::Malicious => {
-            query_type = QueryType::MaliciousIpa(ipa_query_config.clone())
-        }
-    };
-
-    let query_config = QueryConfig {
-        field_type: args.input.field,
-        query_type,
-    };
-    let query_id = helper_clients[0].create_query(query_config).await.unwrap();
-    let input_rows = input.iter::<TestRawDataRecord>().collect::<Vec<_>>();
-    let query_size = input_rows.len();
-
-    let expected = {
-        let mut r = ipa_in_the_clear(
-            &input_rows,
-            ipa_query_config.per_user_credit_cap,
-            ipa_query_config.attribution_window_seconds,
-        );
-
-        // pad the output vector to the max breakdown key, to make sure it is aligned with the MPC results
-        // truncate shouldn't happen unless in_the_clear is badly broken
-        r.resize(
-            usize::try_from(ipa_query_config.max_breakdown_key).unwrap(),
-            0,
-        );
-        r
-    };
-
-    // `key_registries` holds the owned registries. `match_key_encryption` has the key ID and
-    // registry references that we pass to playbook_ipa.
-    let mut key_registries = KeyRegistries::default();
-    let match_key_encryption = if ipa_query_config.plaintext_match_keys {
-        None
-    } else {
-        match key_registries.init_from(network) {
-            mk_enc @ Some(_) => mk_enc,
-            None => panic!(
-                "match key encryption was requested, but one or more helpers is missing a public key"
-            ),
-        }
-    };
-
-    let mpc_time = Instant::now();
-    let actual = match args.input.field {
-        FieldType::Fp31 => {
-            playbook_ipa::<Fp31, MatchKey, BreakdownKey, _>(
-                &input_rows,
-                &helper_clients,
-                query_id,
-                match_key_encryption,
-            )
-            .await
-        }
-        FieldType::Fp32BitPrime => {
-            playbook_ipa::<Fp32BitPrime, MatchKey, BreakdownKey, _>(
-                &input_rows,
-                &helper_clients,
-                query_id,
-                match_key_encryption,
-            )
-            .await
-        }
-    };
-    tracing::info!(
-        "{m:?} IPA for {q} records took {t:?}",
-        m = ipa_query_config,
-        q = query_size,
-        t = mpc_time.elapsed()
-    );
-
-    validate(expected, actual)
 }
 
 async fn multiply_in_field<F: Field>(
