+++ tests/helper_networks.rs
@@ -19,6 +19,7 @@ pub mod tempdir;
 
 const HELPER_BIN: &str = env!("CARGO_BIN_EXE_helper");
 const TEST_MPC_BIN: &str = env!("CARGO_BIN_EXE_test_mpc");
+const TEST_RC_BIN: &str = env!("CARGO_BIN_EXE_report_collector");
 
 trait UnwrapStatusExt {
     fn unwrap_status(self);
@@ -205,7 +206,7 @@ fn test_ipa(mode: IpaSecurityModel, https: bool) {
 
     // Gen inputs
     let inputs_file = dir.path().join("ipa_inputs.txt");
-    let mut command = Command::new(TEST_MPC_BIN);
+    let mut command = Command::new(TEST_RC_BIN);
     command
         .arg("gen-ipa-inputs")
         .args(["--count", "10"])
@@ -216,7 +217,7 @@ fn test_ipa(mode: IpaSecurityModel, https: bool) {
     command.status().unwrap_status();
 
     // Run IPA
-    let mut command = Command::new(TEST_MPC_BIN);
+    let mut command = Command::new(TEST_RC_BIN);
     command
         .args(["--network".into(), dir.path().join("network.toml")])
         .args(["--input-file".as_ref(), inputs_file.as_os_str()])
